Move constraint classes to the separate directory
