+++ sawatabi/model/constraint/abstract_constraint.py
